chore: dissolve `currency` file
