+++ packages/shared/components/popups/ActivityDetailsPopup.svelte
@@ -14,8 +14,7 @@
         selectedAccountActivities,
     } from '@core/wallet'
     import { activeProfile, checkActiveProfileAuth } from '@core/profile'
-    import { currencies, exchangeRates } from '@lib/currency'
-    import { CurrencyTypes } from 'shared/lib/typings/currency'
+    import { currencies, Currency, exchangeRates } from '@core/utils'
     import { setClipboard } from '@core/utils'
     import { truncateString } from '@core/utils'
     import { closePopup, openPopup } from '@auxiliary/popup'
@@ -48,7 +47,7 @@
             transactionTime: activity.time,
             inclusionState: activity.inclusionState,
             formattedFiatValue: activity.getFiatAmount(
-                $currencies[CurrencyTypes.USD],
+                $currencies[Currency.USD],
                 $exchangeRates[$activeProfile?.settings?.currency]
             ),
         }
