+++ packages/shared/components/popups/send/SendConfirmationPopup.svelte
@@ -33,9 +33,9 @@
         newTransactionDetails,
         updateNewTransactionDetails,
     } from '@core/wallet'
-    import { convertToFiat, currencies, exchangeRates, formatCurrency } from '@lib/currency'
+    import { formatCurrency } from '@core/i18n'
+    import { currencies, Currency, exchangeRates, miotaToFiat } from '@core/utils'
     import { closePopup, openPopup } from '@auxiliary/popup'
-    import { CurrencyTypes } from '@lib/typings/currency'
     import { BaseError } from '@core/error'
     import { ledgerPreparedOutput } from '@core/ledger'
 
@@ -69,11 +69,7 @@
 
     $: formattedFiatValue =
         formatCurrency(
-            convertToFiat(
-                Big(rawAmount),
-                $currencies[CurrencyTypes.USD],
-                $exchangeRates[$activeProfile?.settings?.currency]
-            )
+            miotaToFiat(Big(rawAmount), $currencies[Currency.USD], $exchangeRates[$activeProfile?.settings?.currency])
         ) || ''
 
     $: transactionDetails = {
