+++ packages/shared/lib/core/i18n/constants/decimal-separators.constant.ts
@@ -1,0 +1,1 @@
+export const DECIMAL_SEPARATORS = ['.', ',']
