+++ packages/shared/lib/core/i18n/constants/default-locale-options.constant.ts
@@ -1,0 +1,9 @@
+import { LocaleOptions } from '../types'
+
+export const DEFAULT_LOCALE_OPTIONS: LocaleOptions = {
+    fallbackLocale: 'en',
+    initialLocale: null,
+    loadingDelay: 200,
+    formats: {},
+    warnOnMissingMessages: true,
+}
