+++ packages/shared/lib/core/i18n/constants/index.ts
@@ -1,1 +1,3 @@
-export * from './locale'
+export * from './decimal-separators.constant'
+export * from './default-locale-options.constant'
+export * from './supported-locales.constant'
