+++ packages/shared/lib/core/i18n/index.ts
@@ -2,6 +2,7 @@
 export * from './constants'
 export * from './stores'
 export * from './types'
+export * from './utils'
 
 // FILES
 export * from './i18n'
