+++ packages/shared/lib/core/i18n/utils/replaceCurrencyDecimal.ts
@@ -1,0 +1,5 @@
+import { getDecimalSeparator } from './getDecimalSeparator'
+
+export function replaceCurrencyDecimal(value: string, currency: string | undefined = undefined): string {
+    return value.replace('.', getDecimalSeparator(currency))
+}
