+++ packages/shared/lib/core/profile/constants/default-active-profile-values.constant.ts
@@ -1,5 +1,6 @@
 import { NetworkProtocol, NetworkType } from '@core/network'
-import { AvailableExchangeRates } from '@lib/typings/currency'
+import { ExchangeRate } from '@core/utils'
+
 import { ProfileType } from '../enums'
 import { IPersistedProfile } from '../interfaces'
 import { PROFILE_VERSION } from './profile-version.constant'
@@ -13,7 +14,7 @@ export const DEFAULT_ACTIVE_PROFILE_VALUE: IPersistedProfile = {
     networkType: NetworkType?.Mainnet,
     lastStrongholdBackupTime: new Date(),
     settings: {
-        currency: AvailableExchangeRates.USD,
+        currency: ExchangeRate.USD,
         lockScreenTimeoutInMinutes: 5,
         hideNetworkStatistics: true,
     },
