+++ packages/shared/lib/core/utils/constants/default-exchange-rates.constant.ts
@@ -1,0 +1,38 @@
+import { ExchangeRate } from '../enums'
+import { ExchangeRates } from '../types'
+
+export const DEFAULT_EXCHANGE_RATES: ExchangeRates = {
+    [ExchangeRate.AUD]: 1,
+    [ExchangeRate.BGN]: 1,
+    [ExchangeRate.BRL]: 1,
+    [ExchangeRate.CAD]: 1,
+    [ExchangeRate.CHF]: 1,
+    [ExchangeRate.CNY]: 1,
+    [ExchangeRate.CZK]: 1,
+    [ExchangeRate.DKK]: 1,
+    [ExchangeRate.EUR]: 1,
+    [ExchangeRate.GBP]: 1,
+    [ExchangeRate.HKD]: 1,
+    [ExchangeRate.HRK]: 1,
+    [ExchangeRate.HUF]: 1,
+    [ExchangeRate.IDR]: 1,
+    [ExchangeRate.ILS]: 1,
+    [ExchangeRate.INR]: 1,
+    [ExchangeRate.ISK]: 1,
+    [ExchangeRate.JPY]: 1,
+    [ExchangeRate.KRW]: 1,
+    [ExchangeRate.MXN]: 1,
+    [ExchangeRate.MYR]: 1,
+    [ExchangeRate.NOK]: 1,
+    [ExchangeRate.NZD]: 1,
+    [ExchangeRate.PHP]: 1,
+    [ExchangeRate.PLN]: 1,
+    [ExchangeRate.RON]: 1,
+    [ExchangeRate.RUB]: 1,
+    [ExchangeRate.SEK]: 1,
+    [ExchangeRate.SGD]: 1,
+    [ExchangeRate.THB]: 1,
+    [ExchangeRate.TRY]: 1,
+    [ExchangeRate.USD]: 1,
+    [ExchangeRate.ZAR]: 1,
+}
