+++ packages/shared/lib/core/utils/constants/index.ts
@@ -1,4 +1,5 @@
 export * from './bech32-address-length.constant'
+export * from './default-exchange-rates.constant'
 export * from './hex.constants'
 export * from './iota-unit-map.constant'
 export * from './max-number-of-iotas.constant'
