+++ packages/shared/lib/core/utils/enums/currency.enum.ts
@@ -1,0 +1,7 @@
+export enum Currency {
+    IOTA = 'iota',
+    BTC = 'btc',
+    ETH = 'eth',
+    EUR = 'eur',
+    USD = 'usd',
+}
