+++ packages/shared/lib/core/utils/enums/index.ts
@@ -1,2 +1,4 @@
+export * from './currency.enum'
+export * from './exchange-rate.enum'
 export * from './expiration-time.enum'
 export * from './iota-unit.enum'
