+++ packages/shared/lib/core/utils/index.ts
@@ -1,6 +1,7 @@
 export * from './constants'
 export * from './enums'
 export * from './interfaces'
+export * from './stores'
 export * from './types'
 
 export * from './convert'
