+++ packages/shared/lib/core/utils/stores/currencies.store.ts
@@ -1,0 +1,5 @@
+import { writable } from 'svelte/store'
+
+import { Currencies } from '../types'
+
+export const currencies = writable<Currencies>({} as Currencies)
