+++ packages/shared/lib/core/utils/stores/exchange-rates.store.ts
@@ -1,0 +1,6 @@
+import { writable } from 'svelte/store'
+
+import { DEFAULT_EXCHANGE_RATES } from '../constants'
+import { ExchangeRates } from '../types'
+
+export const exchangeRates = writable<ExchangeRates>(DEFAULT_EXCHANGE_RATES)
