+++ packages/shared/lib/core/utils/stores/index.ts
@@ -1,0 +1,2 @@
+export * from './currencies.store'
+export * from './exchange-rates.store'
