+++ packages/shared/lib/core/utils/types/currencies.type.ts
@@ -1,0 +1,5 @@
+import { Currency } from '../enums'
+
+export type Currencies = {
+    [key in Currency]: number
+}
