+++ packages/shared/lib/core/utils/types/exchange-rates.type.ts
@@ -1,0 +1,5 @@
+import { ExchangeRate } from '../enums'
+
+export type ExchangeRates = {
+    [key in ExchangeRate]: number
+}
