+++ packages/shared/lib/core/utils/types/index.ts
@@ -1,3 +1,5 @@
+export * from './currencies.type'
 export * from './duration.type'
+export * from './exchange-rates.type'
 export * from './iota-unit-map.type'
 export * from './values-of.type'
