+++ packages/shared/lib/typings/index.ts
@@ -1,5 +1,4 @@
-import * as currency from './currency'
 import * as events from './events'
 import * as message from './message'
 
-export { currency, events, message }
+export { events, message }
